alx_travel_app/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.14,<4.0
django-filter>=23.0,<24.0
mysqlclient>=2.1.0
pillow>=9.0.0
requests>=2.28.0
django-cors-headers>=4.0,<5.0
drf-yasg>=1.21,<2.0